Move data access to central location. Renamed some things. Organized some imports (#4271)
