+++ client-react/src/pages/app/functions/integrate/BindingPanel/BindingEditor.tsx
