+++ client-react/src/pages/app/functions/integrate/BindingsDiagram/TriggerBindingCard.tsx
@@ -1,20 +1,20 @@
-import React, { useContext } from 'react';
-import BindingCard, { BindingCardChildProps, editExisting, emptyList } from './BindingCard';
-import { ReactComponent as PowerSvg } from '../../../../../images/Common/power.svg';
-import { useTranslation } from 'react-i18next';
-import { BindingInfo } from '../../../../../models/functions/function-binding';
-import { getBindingConfigDirection } from '../binding-editor/BindingEditor';
-import { BindingConfigDirection } from '../../../../../models/functions/bindings-config';
-import PortalCommunicator from '../../../../../portal-communicator';
-import { ArmObj } from '../../../../../models/arm-obj';
-import { FunctionInfo } from '../../../../../models/functions/function-info';
 import i18next from 'i18next';
-import { BindingEditorContextInfo, BindingEditorContext } from '../FunctionIntegrate';
-import { ThemeExtended } from '../../../../../theme/SemanticColorsExtended';
-import { listStyle } from './BindingDiagram.styles';
 import { Link } from 'office-ui-fabric-react';
-import { ThemeContext } from '../../../../../ThemeContext';
+import React, { useContext } from 'react';
+import { useTranslation } from 'react-i18next';
+import { ReactComponent as PowerSvg } from '../../../../../images/Common/power.svg';
+import { ArmObj } from '../../../../../models/arm-obj';
+import { BindingConfigDirection } from '../../../../../models/functions/bindings-config';
+import { BindingInfo } from '../../../../../models/functions/function-binding';
+import { FunctionInfo } from '../../../../../models/functions/function-info';
+import PortalCommunicator from '../../../../../portal-communicator';
 import { PortalContext } from '../../../../../PortalContext';
+import { ThemeExtended } from '../../../../../theme/SemanticColorsExtended';
+import { ThemeContext } from '../../../../../ThemeContext';
+import { getBindingConfigDirection } from '../BindingPanel/BindingEditor';
+import { BindingEditorContext, BindingEditorContextInfo } from '../FunctionIntegrate';
+import BindingCard, { BindingCardChildProps, editExisting, emptyList } from './BindingCard';
+import { listStyle } from './BindingDiagram.styles';
 
 const TriggerBindingCard: React.SFC<BindingCardChildProps> = props => {
   const { functionInfo } = props;
